edit/delete sorted

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -5,6 +5,8 @@ import ImageHoster.model.Tag;
 import ImageHoster.model.User;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
+import ImageHoster.service.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -94,13 +96,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, 
+    						@RequestParam("userId") Integer userId,
+    						Model model) {
         Image image = imageService.getImage(imageId);
-
+        
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        if(userId == image.getUser().getId()) {
+        	model.addAttribute("editError", false);
+        	return "images/edit";	
+        }
+        model.addAttribute("editError", true);
+        return "images/image";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -142,9 +151,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,
+    								@RequestParam(name = "userId") Integer userId, Model model) {
+    	Image image = imageService.getImage(imageId);
+    	String tags = convertTagsToString(image.getTags());
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+        
+    	if(userId == image.getUser().getId()) {
+        	model.addAttribute("deleteError", false);
+        	imageService.deleteImage(imageId);
+            return "redirect:/images";	
+        }
+        model.addAttribute("deleteError", true);
+        return "images/image";
+    	
     }
